requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Deprecated==1.2.14
fastapi==0.114.2
filelock==3.13.1
fiona==1.10.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.6.0
geopandas==0.14.4
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
idna==3.8
Jinja2==3.1.3
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
networkx==2.8.8
numpy==2.1.1
opt-einsum==3.3.0
optree==0.12.1
osmnx==1.9.4
overpy==0.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
psutil==6.0.0
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python_tsp==0.5.0
pytz==2024.2
requests==2.32.3
rich==13.8.1
scikit-learn==1.5.2
scipy==1.14.1
setuptools==74.1.2
shapely==2.0.6
simplekml==1.3.6
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
sympy==1.12
tabulate==0.8.10
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.4.1+cpu
torch-geometric==2.6.0
tqdm==4.66.5
tsplib95==0.7.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
yarl==1.11.1
